# src/main/resources/swagger/service.yaml
...
openapi: 3.0.2
info:
  title: Reasoner
  description: Reasoner is engine for OntoWeb project. It provides Rest API for back-office application and for client queries.
  license:
    name: © Nikolas Lada
  version: 0.2.1-oas3
servers:
- url: /
paths:
...
  /user:
    post:
      summary: Create a User
      description: Creates a new User.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserRequest'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        "400":
          description: HTTP Client Error 4xx
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/{by}/{value}:
    get:
      summary: Get the User Detail by id or login.
      parameters:
      - name: by
        in: path
        description: id|login
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: value
        in: path
        description: Numeric id or login.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        "400":
          description: HTTP Client Error 4xx
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: HTTP Client Error 4xx
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      title: Error
      required:
      - code
      - messageList
      - timestamp
      type: object
      properties:
        code:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
          example: 2020-01-13T12:00:00Z
        messageList:
          type: array
          description: List of error codes and error messages.
          items:
            $ref: '#/components/schemas/ErrorMessage'
      example:
        code: 400
        timestamp: 2020-01-13T12:00:00.000Z
        messageList:
        - code: notValidClass
          message: Class is not valid!
    ErrorMessage:
      title: ErrorMessage
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      example:
        code: notValidClass
        message: Class is not valid!
...
    NewUserRequest:
      title: NewUserRequest
      type: object
      properties:
        firstName:
          type: string
        surname:
          type: string
        login:
          type: string
        password:
          type: string
    UserDetailResponse:
      title: UserDetailResponse
      required:
      - createdAt
      - firstName
      - login
      - surname
      - updatedAt
      type: object
      properties:
        id:
          type: integer
          format: int32
        login:
          type: string
        firstName:
          type: string
        surname:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...